Create Team 6
Performing of VAT technique for determining the potential cluster and the possible number of clusters in speech data and pictorial representation of clusters performed on various datasets and Validating it by “K-Means”, ”MST” and Normalized Mutual Information (NMI).The proposed work is to enable solutions of two key issues, these are assessment of number speakers and effective speech clustering results by analyzing the voice models of speakers data. In this, we used one of the efficient statistical modeling technique viz., Gaussian Mixture Model (hereinafter, GMM) for constructing models of speech segments based on the extracted features of speakers data. In speech cluster analysis, similar features of GMM mean vi supervectors for the speech segments are derived using distance metrics, in specific, as per our observation, Cosine based similarity is greatly succeeded for speech data to derive the similarity features